Goals can be moved up and down and that is saved in the db.
